feat: Add Retrofit integration for full-stack connectivity
- Added Retrofit + OkHttp network layer
- Created API services (Auth, Transaction, Account, Analytics, SMS)
- Created DTOs matching backend responses
- Added TokenManager for JWT storage
- Added AuthInterceptor for automatic token injection
- Fixed domain model alignment
- Connected AuthViewModel to backend API

+++ android/app/src/main/java/com/praveen/expensetracker/domain/usecase/GetDashboardDataUseCase.kt
@@ -28,8 +28,8 @@ class GetDashboardDataUseCase @Inject constructor(
         val totalExpense = repository.getTotalExpense(startOfMonth, endOfMonth)
         val safeToSpend = (monthlyBudget - totalExpense).coerceAtLeast(0.0)
         val budgetUsedPercentage = if (monthlyBudget > 0) {
-            (totalExpense / monthlyBudget * 100).toFloat()
-        } else 0f
+            (totalExpense / monthlyBudget * 100)
+        } else 0.0
         
         val recentTransactions = repository.getRecentTransactions(10).first()
